refactor(infra): remove storage shim and align domain infra docs
